+++ CHANGES.md
@@ -1,11 +1,12 @@
-Elektron Change Log
+Elektrum Change Log
 ================
 
 0.32 (unreleased)
 -----------------
 
 - [#57] Implement media upload.
-
+- [#58] Automate provisioning of different environments.
+- Renamed project from `elektron` to `elektrum` because the domain name was available in the `.dev` TLD as well as `.photos`.
 
 0.31 (2019-07-25)
 -----------------